+++ envoy-control-tests/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/config/service/UpstreamService.kt
@@ -1,0 +1,8 @@
+package pl.allegro.tech.servicemesh.envoycontrol.config.service
+
+import pl.allegro.tech.servicemesh.envoycontrol.config.envoy.ResponseWithBody
+
+interface UpstreamService {
+    fun id(): String
+    fun isSourceOf(response: ResponseWithBody): Boolean
+}