+++ envoy-control-tests/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/routing/RoutingHeadersTest.kt
@@ -29,15 +29,19 @@ class RoutingHeadersTest : TestBase(echoService, envoy) {
         val envoy = EnvoyExtension(envoyControl, config = RandomConfigFile.copy(configOverride = proxySettings))
     }
 
-    @Test
-    fun `should add correct x-service-tag-preference header to upstream request`() {
-        // given
+    private fun registerServices() {
         listOf("echo", "echo-disabled", "echo-one-tag", "echo-no-tag").forEach { service ->
             consul.server.operations.registerService(name = service, extension = echoService, tags = allTags)
         }
         listOf("echo", "echo-disabled", "echo-one-tag", "echo-no-tag").forEach { service ->
             waitForEndpointReady(service, echoService, envoy)
         }
+    }
+
+    @Test
+    fun `should add correct x-service-tag-preference header to upstream request`() {
+        // given
+        registerServices()
 
         // when
         val echoResponse = envoy.egressOperations.callService("echo").asHttpsEchoResponse()
@@ -60,8 +64,31 @@ class RoutingHeadersTest : TestBase(echoService, envoy) {
     }
 
     @Test
+    fun `should not override service-tag preference header already set in the request`() {
+        // given
+        registerServices()
+
+        // when
+        val echoResponse = envoy.egressOperations.callService(
+            service = "echo",
+            headers = mapOf("x-service-tag-preference" to "custom|ipsum|lorem")
+        ).asHttpsEchoResponse()
+
+        // then
+        assertThat(echoResponse).isOk()
+        assertThat(echoResponse.requestHeaders).containsEntry("x-service-tag-preference", "custom|ipsum|lorem")
+    }
+
+    @Test
     fun `should add upstream service tags to a response`() =
-        upstreamServiceTagsInResponseTest { echoResponse ->
+        upstreamServiceTagsInResponseTest(service = "echo") { echoResponse ->
+            // then
+            assertThat(echoResponse.headers("x-envoy-upstream-service-tags")).isEqualTo(listOf("""["ipsum","lorem","one"]"""))
+        }
+
+    @Test
+    fun `should add upstream service tags to a response even if autoServiceTag is disabled`() =
+        upstreamServiceTagsInResponseTest(service = "echo-disabled") { echoResponse ->
             // then
             assertThat(echoResponse.headers("x-envoy-upstream-service-tags")).isEqualTo(listOf("""["ipsum","lorem","one"]"""))
         }
@@ -83,9 +110,10 @@ class UpstreamTagHeadersDisabledTest : TestBase(echoService, envoy) {
     }
 
     @Test
-    fun `should not add upstream service tags to a response`() = upstreamServiceTagsInResponseTest { echoResponse ->
-        assertThat(echoResponse.headers("x-envoy-upstream-service-tags")).isEmpty()
-    }
+    fun `should not add upstream service tags to a response`() =
+        upstreamServiceTagsInResponseTest(service = "echo") { echoResponse ->
+            assertThat(echoResponse.headers("x-envoy-upstream-service-tags")).isEmpty()
+        }
 }
 
 abstract class TestBase(private val echoService: ServiceExtension<*>, private val envoy: EnvoyExtension) {
@@ -126,13 +154,13 @@ abstract class TestBase(private val echoService: ServiceExtension<*>, private va
         val allTags = listOf("ipsum", "lorem", "one")
     }
 
-    protected fun upstreamServiceTagsInResponseTest(assertions: (echoResponse: Response) -> Unit) {
+    protected fun upstreamServiceTagsInResponseTest(service: String, assertions: (echoResponse: Response) -> Unit) {
         // given
-        consul.server.operations.registerService(name = "echo", extension = echoService, tags = allTags)
-        waitForEndpointReady("echo", echoService, envoy)
+        consul.server.operations.registerService(name = service, extension = echoService, tags = allTags)
+        waitForEndpointReady(service, echoService, envoy)
 
         // when
-        val echoResponse = envoy.egressOperations.callService("echo")
+        val echoResponse = envoy.egressOperations.callService(service)
 
         // then
         assertions(echoResponse)