+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 Lists all changes with user impact.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [0.22.12]
+### Changed
+- add service tag preference routing
+- optimize auto service tag routing (don't send config to envoys if disabled, reduce number of metadata ser per route)
+- fix and refactor e2e tests
+
 ## [0.22.11]
 ### Changed
 - Implemented handling of initialResourcesVersions in DeltaRequest for ADS